# Slightly refactor and pre-approve when possible