… packages/protocol/test/common/onboarding.ts